example/cluster/worker/requirements.txt
queuerPy==0.4.0
openai-whisper==20231117